Multiple Css chagnes to menu, header, and create issue
