+++ client/src/components/Menu/Menu.module.css
@@ -4,7 +4,7 @@
   bottom: 0;
   display: grid;
   grid-template-columns: 200px;
-  grid-template-rows: repeat(7, 0.5fr);
+  grid-template-rows: repeat(8, 0.5fr);
 }
 
 .NavBarContainer button {
@@ -13,4 +13,11 @@
   height: 3rem;
   width: 11rem;
   border-radius: 32px;
+}
+
+.conditionalRenderP {
+  text-align: center;
+  margin: auto;
+  font-size: 22px;
+  cursor: pointer;
 }
