+++ client/src/components/header/header.module.css
@@ -15,12 +15,6 @@
   background-color: #34435e;
 }
 
-.conditionalRenderP {
-  margin: auto;
-  font-size: 14px;
-  cursor: pointer;
-}
-
 .headerContainer div {
   margin: auto;
 }
